model-router/requirements.txt
openai 
python-dotenv 
streamlit
asyncio